pyproject.toml
[tool.poetry]
name = "nasy_stats_bot"
version = "0.1.0"
description = "Nasy's Telegram Stats Bot"
authors = ["Nasy <nasyxx+git@gmail.com>"]
license = "GPL-3.0+"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^12.4.2"
pendulum = "^2.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
